requirements.txt
django>=2
pip
pycrypto
setuptools
wheel